Start migration from Sidekiq::Worker to Sidekiq::Job
* Provide Sidekiq::Worker alias for backwards compat
* Rename API's Sidekiq::Job to Sidekiq::JobRecord
* Adjust internals to use Sidekiq::JobRecord
