+++ lib/generators/sidekiq/worker_generator.rb
@@ -3,16 +3,16 @@ require "rails/generators/named_base"
 module Sidekiq
   module Generators # :nodoc:
     class WorkerGenerator < ::Rails::Generators::NamedBase # :nodoc:
-      desc "This generator creates a Sidekiq Worker in app/workers and a corresponding test"
+      desc "This generator creates a Sidekiq::Job in app/workers and a corresponding test"
 
-      check_class_collision suffix: "Worker"
+      check_class_collision suffix: "Job"
 
       def self.default_generator_root
         File.dirname(__FILE__)
       end
 
       def create_worker_file
-        template "worker.rb.erb", File.join("app/workers", class_path, "#{file_name}_worker.rb")
+        template "worker.rb.erb", File.join("app/workers", class_path, "#{file_name}_job.rb")
       end
 
       def create_test_file
@@ -31,7 +31,7 @@ module Sidekiq
         template_file = File.join(
           "spec/workers",
           class_path,
-          "#{file_name}_worker_spec.rb"
+          "#{file_name}_job_spec.rb"
         )
         template "worker_spec.rb.erb", template_file
       end
@@ -40,13 +40,13 @@ module Sidekiq
         template_file = File.join(
           "test/workers",
           class_path,
-          "#{file_name}_worker_test.rb"
+          "#{file_name}_job_test.rb"
         )
         template "worker_test.rb.erb", template_file
       end
 
       def file_name
-        @_file_name ||= super.sub(/_?worker\z/i, "")
+        @_file_name ||= super.sub(/_?job\z/i, "")
       end
 
       def test_framework
