+++ lib/sidekiq/extensions/class_methods.rb
@@ -13,7 +13,7 @@ module Sidekiq
     #   Wikipedia.delay.download_changes_for(Date.today)
     #
     class DelayedClass
-      include Sidekiq::Worker
+      include Sidekiq::Job
 
       def perform(yml)
         (target, method_name, args) = YAML.load(yml)
