+++ test/test_rails.rb
@@ -9,19 +9,19 @@ describe 'ActiveJob' do
     # need to force this since we aren't booting a Rails app
     ActiveJob::Base.queue_adapter = :sidekiq
     ActiveJob::Base.logger = nil
-    ActiveJob::Base.send(:include, ::Sidekiq::Worker::Options) unless ActiveJob::Base.respond_to?(:sidekiq_options)
+    ActiveJob::Base.send(:include, ::Sidekiq::Job::Options) unless ActiveJob::Base.respond_to?(:sidekiq_options)
   end
 
-  it 'does not allow Sidekiq::Worker in AJ::Base classes' do
+  it 'does not allow Sidekiq::Job in AJ::Base classes' do
     ex = assert_raises ArgumentError do
       Class.new(ActiveJob::Base) do
-        include Sidekiq::Worker
+        include Sidekiq::Job
       end
     end
-    assert_includes ex.message, "Sidekiq::Worker cannot be included"
+    assert_includes ex.message, "Sidekiq::Job cannot be included"
   end
 
-  it 'loads Sidekiq::Worker::Options in AJ::Base classes' do
+  it 'loads Sidekiq::Job::Options in AJ::Base classes' do
     aj = Class.new(ActiveJob::Base) do
       queue_as :bar
       sidekiq_options retry: 4, queue: 'foo', backtrace: 5
